chore: improve exception for type validation (#3759)
change `InvalidType` error to more accurate `TypeMismatch`
