+++ tests/functional/codegen/features/test_assignment.py
@@ -207,7 +207,7 @@ def foo2() -> uint256:
     x += 1
     return x
 """
-    assert_compile_failed(lambda: get_contract_with_gas_estimation(code), InvalidType)
+    assert_compile_failed(lambda: get_contract_with_gas_estimation(code), TypeMismatch)
 
 
 def test_invalid_uin256_assignment_calculate_literals(get_contract_with_gas_estimation):
