+++ tests/functional/syntax/test_addmulmod.py
@@ -1,7 +1,7 @@
 import pytest
 
 from vyper import compile_code
-from vyper.exceptions import InvalidType
+from vyper.exceptions import TypeMismatch
 
 fail_list = [
     (  # bad AST nodes given as arguments
@@ -10,7 +10,7 @@ fail_list = [
 def foo() -> uint256:
     return uint256_addmod(1.1, 1.2, 3.0)
     """,
-        InvalidType,
+        TypeMismatch,
     ),
     (  # bad AST nodes given as arguments
         """
@@ -18,7 +18,7 @@ def foo() -> uint256:
 def foo() -> uint256:
     return uint256_mulmod(1.1, 1.2, 3.0)
     """,
-        InvalidType,
+        TypeMismatch,
     ),
 ]
