+++ tests/functional/syntax/test_concat.py
@@ -1,7 +1,7 @@
 import pytest
 
 from vyper import compiler
-from vyper.exceptions import ArgumentException, InvalidType, TypeMismatch
+from vyper.exceptions import ArgumentException, TypeMismatch
 
 fail_list = [
     (
@@ -18,7 +18,7 @@ def cat(i1: Bytes[10], i2: Bytes[30]) -> Bytes[40]:
 def cat(i1: Bytes[10], i2: Bytes[30]) -> Bytes[40]:
     return concat(i1, 5)
     """,
-        InvalidType,
+        TypeMismatch,
     ),
     (
         """
