+++ tests/functional/syntax/test_logging.py
@@ -1,7 +1,7 @@
 import pytest
 
 from vyper import compiler
-from vyper.exceptions import InvalidType, StructureException, TypeMismatch
+from vyper.exceptions import StructureException, TypeMismatch
 
 fail_list = [
     """
@@ -23,8 +23,7 @@ def foo():
     x: decimal[4] = [0.0, 0.0, 0.0, 0.0]
     log Bar(x)
     """,
-    (
-        """
+    """
 event Test:
     n: uint256
 
@@ -32,19 +31,13 @@ event Test:
 def test():
     log Test(-7)
    """,
-        InvalidType,
-    ),
 ]
 
 
 @pytest.mark.parametrize("bad_code", fail_list)
 def test_logging_fail(bad_code):
-    if isinstance(bad_code, tuple):
-        with pytest.raises(bad_code[1]):
-            compiler.compile_code(bad_code[0])
-    else:
-        with pytest.raises(TypeMismatch):
-            compiler.compile_code(bad_code)
+    with pytest.raises(TypeMismatch):
+        compiler.compile_code(bad_code)
 
 
 @pytest.mark.parametrize("mutability", ["@pure", "@view"])
