+++ tests/unit/cli/vyper_json/test_compile_json.py
@@ -12,7 +12,7 @@ from vyper.cli.vyper_json import (
 )
 from vyper.compiler import OUTPUT_FORMATS, compile_code, compile_from_file_input
 from vyper.compiler.input_bundle import JSONInputBundle
-from vyper.exceptions import InvalidType, JSONError, SyntaxException
+from vyper.exceptions import JSONError, SyntaxException, TypeMismatch
 
 FOO_CODE = """
 import contracts.ibar as IBar
@@ -244,7 +244,7 @@ def test_exc_handler_to_dict_syntax(input_json):
 
 def test_exc_handler_raises_compiler(input_json):
     input_json["sources"]["badcode.vy"] = {"content": BAD_COMPILER_CODE}
-    with pytest.raises(InvalidType):
+    with pytest.raises(TypeMismatch):
         compile_json(input_json)
 
 
@@ -256,7 +256,7 @@ def test_exc_handler_to_dict_compiler(input_json):
     assert len(result["errors"]) == 1
     error = result["errors"][0]
     assert error["component"] == "compiler"
-    assert error["type"] == "InvalidType"
+    assert error["type"] == "TypeMismatch"
 
 
 def test_source_ids_increment(input_json):
